Force storage.file_overwrite to True
This fixes uploading a static file with the same name twice resulting in
a second file with a randomized suffix in the name, and of course the
first one never to be updated to the newer version.

This is a bit related to #30 and basically has the same fix. It's a bit
surprising that this never was a problem before, but that's probably
because, in contrary to the setting from #30, that default in
django-storages for this setting is the one we need. This also means
that the impact of this change is a lot smaller. It only affects
projects which have manually set this setting to False, most likely to
prevent user uploads from overwriting each other.